pyproject.toml
[tool.poetry]
name = "openbci2"
version = "0.1.0"
description = ""
authors = ["Ray Chen <raychen256@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
brainflow = "^5.12.1"
setuptools = "^69.1.1"
matplotlib = "^3.8.3"
ipykernel = "^6.29.3"
sounddevice = "^0.4.6"
soundfile = "^0.12.1"
python-dotenv = "^1.0.1"
flask = "^3.0.2"
werkzeug = "^3.0.1"
pandas = "^2.2.1"
scikit-learn = "^1.4.1.post1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
